added a database that a user can now add information to via a form on the albums page, and view their updated page
