+++ README.md
@@ -11,7 +11,7 @@ Sharina Stubbs
 
 ## Run this application
 1. Git clone this repo into your local system
-2. Have songr Build the project
+2. Run `.gradlew bootrun`
 3. Go to localhost:8080 to see the splash page
 4. Go to localhost:8080/capitalize/your-word-or-sentence-here to change your word or sentence to uppercase.
 5. Go to localhost:8080/albums to see three albums and their data
