+++ src/main/resources/application.properties
@@ -1,1 +1,2 @@
-
+spring.datasource.url=jdbc:postgresql://localhost:5432/albums
+#spring.jpa.hibernate.ddl-auto=create
